[UPDATE] 화상 효과에 트윈 적용

+++ Assets/01.Scripts/Ability/FlameAbility.cs
@@ -3,9 +3,9 @@ using UnityEngine;
 
 public class FlameAbility : IAbility
 {
-    private System.Action<int> burn;
+    private System.Action<float> burn;
 
-    public FlameAbility(System.Action<int> burn)
+    public FlameAbility(System.Action<float> burn)
     {
         this.burn = burn;
     }
@@ -15,7 +15,7 @@ public class FlameAbility : IAbility
         for (int i = 0; i < info.hitCount; i++)
         {
             var monster = activeMonsters[i];
-            burn?.Invoke((int)info.duration);
+            burn?.Invoke(info.duration);
             monster.GetDamage(info.removalCount);
 
             var gObj = Object.Instantiate(effect, monster.transform.position, Quaternion.identity);
